Send P2P connection changed broadcast to system
As of Android 14 QPR2, the WIFI_P2P_CONNECTION_CHANGED_ACTION intent is
not sent to any packages with the MAINLINE_NETWORK_STACK permission,
except the tethering package (for which it runs at a foreground
priority). The system is considered to have this permission, so it's
excluded from receiving the broadcast. One of the effects is that
connecting to a Wi-Fi Display fails.

Fix the system case by sending the broadcast directly to the system.

Bug: 333423704
Change-Id: I004f4706d8bc283b79e7960e7074126e90e012f6